Crate ic_crypto_tls_interfaces

Public interface for a TLS-secured stream

Structs

AllowedClients
A list of allowed TLS peers (and their trusted certificates), which can be All to allow any node to connect.

ClientsEmptyError
Attempted to create an AllowedClients with Some clients but empty nodes and certificates lists.

MalformedPeerCertificateError
The certificate offered by the peer is malformed.

TlsReadHalf
The read half of a stream over a secure connection protected by TLS.

TlsStream
A stream over a secure connection protected by TLS.

TlsWriteHalf
The write half of a stream over a secure connection protected by TLS.

Enums

AuthenticatedPeer
An authenticated Node ID, or an authenticated certificate

Peer
A TLS peer, with identification information (if authenticated)

PeerNotAllowedError
Errors from a TLS handshake due to an unauthorized peer

SomeOrAllNodes
A list of node IDs, or "all nodes"

TlsClientHandshakeError
Errors from a TLS handshake performed as the client. Please refer to the TlsHandshake method for detailed error variant descriptions.

TlsServerHandshakeError
Errors from a TLS handshake performed as the server. Please refer to the TlsHandshake method for detailed error variant descriptions.

Traits

TlsHandshake
Implementors provide methods for transforming TCP streams into TLS stream.