Trait ic_crypto_tls_interfaces::TlsHandshake

pub trait TlsHandshake {
#[must_use]    fn perform_tls_server_handshake<'life0, 'async_trait>(
        &'life0 self, 
        tcp_stream: TcpStream, 
        allowed_clients: AllowedClients, 
        registry_version: RegistryVersion
    ) -> Pin<Box<dyn Future<Output = Result<(TlsStream, AuthenticatedPeer), TlsServerHandshakeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn perform_tls_server_handshake_temp_with_optional_client_auth<'life0, 'async_trait>(
        &'life0 self, 
        tcp_stream: TcpStream, 
        allowed_authenticating_clients: AllowedClients, 
        registry_version: RegistryVersion
    ) -> Pin<Box<dyn Future<Output = Result<(TlsStream, Peer), TlsServerHandshakeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn perform_tls_client_handshake<'life0, 'async_trait>(
        &'life0 self, 
        tcp_stream: TcpStream, 
        server: NodeId, 
        registry_version: RegistryVersion
    ) -> Pin<Box<dyn Future<Output = Result<TlsStream, TlsClientHandshakeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Implementors provide methods for transforming TCP streams into TLS stream.

The TLS streams are returned as trait objects over a trait that does not allow for extracting the secret keys of the underlying TLS session. This is done because directly returning the underlying structs may allow for extraction of the secret session keys.

Required methods

#[must_use]fn perform_tls_server_handshake<'life0, 'async_trait>(
&'life0 self,
tcp_stream: TcpStream,
allowed_clients: AllowedClients,
registry_version: RegistryVersion
) -> Pin<Box<dyn Future<Output = Result<(TlsStream, AuthenticatedPeer), TlsServerHandshakeError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Transforms a TCP stream into a TLS stream by first performing a TLS server handshake and then verifying that the authenticated peer is an allowed client.

For the handshake, the server uses the following configuration:

Minimum protocol version: TLS 1.3

Supported signature algorithms: ed25519

Allowed cipher suites: TLS_AES_128_GCM_SHA256, TLS_AES_256_GCM_SHA384

Client authentication: mandatory, with ed25519 certificate

Maximum number of intermediate CA certificates: 1

To determine whether the peer (that successfully performed the handshake) is an allowed client, the following steps are taken:

Determine the peer's node ID N_claimed from the subject name of the certificate C_handshake that the peer presented during the handshake (and for which the peer therefore knows the private key). If N_claimed is contained in the nodes in allowed_clients, determine the certificate C_registry by querying the registry for the TLS certificate of node with ID N_claimed, and if C_registry is equal to C_handshake, then the peer successfully authenticated as node N_claimed. Otherwise, step 2 is taken.

Compare the root of the certificate chain that the peer presented during the handshake (and for which the peer therefore knows the private key of the chain's leaf certificate) to all the (explicitly allowed) certificates in allowed_clients. If there is a match, then the peer represented by the chain's leaf certificate successfully authenticated.

The given tcp_stream is consumed. If an error is returned, the TCP connection is therefore dropped.

Returns the TLS stream together with the peer that successfully authenticated.

Errors

TlsServerHandshakeError::RegistryError if the registry cannot be accessed.

TlsServerHandshakeError::CertificateNotInRegistry if a certificate that is expected to be in the registry is not found.

TlsServerHandshakeError::MalformedSelfCertificate if the node's own server certificate is malformed.

TlsServerHandshakeError::MalformedClientCertificate if a client certificate corresponding to an client in allowed_clients is malformed.

TlsServerHandshakeError::CreateAcceptorError if there is a problem configuring the server for accepting connections from clients.

TlsServerHandshakeError::HandshakeError if there is an error during the TLS handshake, or the handshake fails.

TlsServerHandshakeError::ClientNotAllowed if the node_id in the subject CN of the client's certificate presented in the handshake is not in allowed_clients, or if the client's certificate presented in the handshake does not exactly match the client's certificate in the registry.

TlsServerHandshakeError::UnauthenticatedClient if the client did not authenticate using a client certificate.

Panics

If the secret key corresponding to the server certificate cannot be found or is malformed in the server's secret key store. Note that this is an error in the setup of the node and registry.

#[must_use]fn perform_tls_server_handshake_temp_with_optional_client_auth<'life0, 'async_trait>(
&'life0 self,
tcp_stream: TcpStream,
allowed_authenticating_clients: AllowedClients,
registry_version: RegistryVersion
) -> Pin<Box<dyn Future<Output = Result<(TlsStream, Peer), TlsServerHandshakeError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

IMPORTANT NODE: This method is temporary. It will be replaced by perform_tls_server_handshake and a new method perform_tls_server_handshake_without_client_auth soon. This method is currently needed to allow connections without knowing if a client performs client authentication.

SECURITY WARNING: The caller of this method is responsible to check if the peer authenticated or not. Only if this method returns Peer::Authenticated it is guaranteed that the client is an allowed client wrt. allowed__authenticating_clients, see below.

Transforms a TCP stream into a TLS stream by performing a TLS server handshake.

For the handshake, the server uses the following configuration:

Minimum protocol version: TLS 1.3

Supported signature algorithms: ed25519

Allowed cipher suites: TLS_AES_128_GCM_SHA256, TLS_AES_256_GCM_SHA384

Client authentication: optional, with ed25519 certificate

Maximum number of intermediate CA certificates: 1

Whenever the TLS handshake fails, this method returns an error.

The given tcp_stream is consumed. If an error is returned, the TCP connection is therefore dropped.

Connections without client authentcation

The client may present a client certificate to authenticate. If it doesn't, Peer:Unauthenticated is returned.

Connections with client authentication

If the client presents a certificate, the TLS handshake only succeeds if the certificate is valid and trusted given the allowed_authenticating_clients. Additionally, after the handshake, to determine whether the peer (that successfully performed the handshake) is an allowed client, the following steps are taken:

Determine the peer's node ID N_claimed from the subject name of the certificate C_handshake that the peer presented during the handshake (and for which the peer therefore knows the private key). If N_claimed is contained in the nodes in allowed_authenticating_clients, determine the certificate C_registry by querying the registry for the TLS certificate of node with ID N_claimed, and if C_registry is equal to C_handshake, then the peer successfully authenticated as node N_claimed. Otherwise, step 2 is taken.

Compare the root of the certificate chain that the peer presented during the handshake (and for which the peer therefore knows the private key of the chain's leaf certificate) to all the (explicitly allowed) certificates in allowed_authenticating_clients. If there is a match, then the peer represented by the chain's leaf certificate successfully authenticated.

If client authentication is successful, the TLS stream together with the peer (Peer::Authenticated) that successfully authenticated is returned.

Errors

TlsServerHandshakeError::RegistryError if the registry cannot be accessed.

TlsServerHandshakeError::CertificateNotInRegistry if a certificate that is expected to be in the registry is not found.

TlsServerHandshakeError::MalformedSelfCertificate if the node's own server certificate is malformed.

TlsServerHandshakeError::MalformedClientCertificate if a client certificate corresponding to an client in allowed_authenticating_clients is malformed.

TlsServerHandshakeError::CreateAcceptorError if there is a problem configuring the server for accepting connections from clients.

TlsServerHandshakeError::HandshakeError if there is an error during the TLS handshake, or the handshake fails.

TlsServerHandshakeError::ClientNotAllowed if the node_id in the subject CN of the client's certificate presented in the handshake is not in allowed_authenticating_clients, or if the client's certificate presented in the handshake does not exactly match the client's certificate in the registry.

Panics

If the secret key corresponding to the server certificate cannot be found or is malformed in the server's secret key store. Note that this is an error in the setup of the node and registry.

#[must_use]fn perform_tls_client_handshake<'life0, 'async_trait>(
&'life0 self,
tcp_stream: TcpStream,
server: NodeId,
registry_version: RegistryVersion
) -> Pin<Box<dyn Future<Output = Result<TlsStream, TlsClientHandshakeError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Transforms a TCP stream into a TLS stream by first performing a TLS client handshake and then verifying that the peer is the given server.

For the handshake, the client uses the following configuration:

Minimum protocol version: TLS 1.3

Supported signature algorithms: ed25519

Allowed cipher suites: TLS_AES_128_GCM_SHA256, TLS_AES_256_GCM_SHA384

Server authentication: mandatory, with ed25519 certificate

To determine whether the peer (that successfully performed the handshake) is the server, the following steps are taken:

Determine the peer's node ID N_claimed from the subject name of the certificate C_handshake that the peer presented during the handshake (and for which the peer therefore knows the private key). Return an error if N_claimed is not the server.

Determine the certificate C_registry by querying the registry for the TLS certificate of node with ID N_claimed. Return an error if the C_registry does not equal C_handshake.

The given tcp_stream is consumed. If an error is returned, the TCP connection is therefore dropped.

Errors

TlsClientHandshakeError::RegistryError if the registry cannot be accessed.

TlsClientHandshakeError::CertificateNotInRegistry if a certificate that is expected to be in the registry is not found.

TlsClientHandshakeError::MalformedSelfCertificate if the node's own client certificate is malformed.

TlsClientHandshakeError::MalformedServerCertificate if the server certificate obtained from the registry (as specified by server) is malformed.

TlsClientHandshakeError::CreateConnectorError if there is a problem configuring the TLS client for connecting to the server.

TlsServerHandshakeError::HandshakeError if there is an error during the TLS handshake, or the handshake fails.

TlsClientHandshakeError::ServerNotAllowed if the node_id in the subject CN of the server's certificate presented in the handshake does not equal server, or if the server's certificate presented in the handshake does not exactly match the server's certificate in the registry.

Panics

If the secret key corresponding to the client certificate cannot be found or is malformed in the client's secret key store. Note that this is an error in the setup of the node and registry.

Loading content...